refactor(providers): move query-provider and theme-provider to src/providers
